Put Python comments in line, move clone of Tomviz
I moved the Python advice to the point of needing it, hopefully helping
people following the sequential instructions. I also moved the Tomviz
clone to right before it is used. I reduced the Python advice to make it
clear that a consistent Python is required, but removed the example of
finding it. We might consider adding some tips, but it is pretty generic
and probably doesn't need to be in our build instructions.