Fix setting of separate color map to preset
Remove unneeded severing of signal/slot connections between the
CentralWidget and the active Module when setActiveColorMapDataSource()
is invoked. One of the severed connections was responsible for setting
the current color map in the HistogramWidget. Because the separate
color map was never being set in the HistogramWidget, changes to the
color map were always being applied to the shared color map. Leaving
these connections in place fixes the problem.

Fixes #1145.